pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scriptorium"
version = "5.0.4"
description = "Intelligent scraper for extracting poems from Wikisource, supporting both online API and offline dump processing modes."
readme = "README.md"
authors = [
  { name = "Sharle", email = "charleskayssieh@gmail.com" },
]
license = { file = "LICENSE" }
requires-python = ">=3.10, <3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Internet :: WWW/HTTP",
    "Framework :: AsyncIO",
]
dependencies = [
    "aiohttp[speedups]>=3.9.0",
    "aiosqlite>=0.19.0",
    "backoff>=2.2.0",
    "pydantic>=2.7.0",
    "tqdm>=4.66.0",
    "mwparserfromhell>=0.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0", # Parseur HTML rapide pour BeautifulSoup
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]

[project.scripts]
scriptorium = "scriptorium.cli:main_cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]